project/requirements.txt
ollama
piper-tts
openai-whisper
moviepy
yt-dlp
google-api-python-client
google-auth
google-auth-oauthlib 
google-auth-httplib2